RF: Demote `save_dataset` to a helper of `save`
We can better tune it if `save` is the only user
